ai-service/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
httpx
aiokafka

# Testing
pytest
pytest-asyncio